增加notify-event线程池配置,实现用户操作的异步处理
