--- src/main/java/org/wj/letsrock/domain/article/model/dto/ReadCountDTO.java
@@ -1,9 +1,0 @@
-package org.wj.letsrock.domain.article.model.dto;
-
-/**
- * @author wujia
- * @description: TODO
- * @createTime: 2025-04-20-14:37
- **/
-public class ReadCountDTO {
-}
