--- src/main/java/org/wj/letsrock/infrastructure/config/AsyncConfig.java
@@ -1,28 +1,0 @@
-package org.wj.letsrock.infrastructure.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.AsyncConfigurer;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-/**
- * @author wujia
- * @description: TODO
- * @createTime: 2025-04-22-20:00
- **/
-@Configuration
-public class AsyncConfig implements AsyncConfigurer {
-    @Override
-    @Bean(name = "taskExecutor")
-    public Executor getAsyncExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(5);        // 核心线程数
-        executor.setMaxPoolSize(10);        // 最大线程数
-        executor.setQueueCapacity(100);     // 队列容量
-        executor.setThreadNamePrefix("DownloadThread-"); // 线程名前缀
-        executor.initialize();
-        return executor;
-    }
-}
