+++ src/main/java/org/wj/letsrock/infrastructure/config/ThreadPoolConfig.java
@@ -1,0 +1,65 @@
+package org.wj.letsrock.infrastructure.config;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.AsyncConfigurer;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.lang.reflect.Method;
+import java.util.concurrent.Executor;
+import java.util.concurrent.ThreadPoolExecutor;
+
+/**
+ * @author wujia
+ * @description: TODO
+ * @createTime: 2025-04-22-20:00
+ **/
+@Configuration
+@Slf4j
+@EnableAsync
+public class ThreadPoolConfig implements AsyncConfigurer {
+    @Override
+    @Bean(name = "imageTransfer")
+    public Executor getAsyncExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(5);        // 核心线程数
+        executor.setMaxPoolSize(10);        // 最大线程数
+        executor.setQueueCapacity(100);     // 队列容量
+        executor.setThreadNamePrefix("DownloadThread-"); // 线程名前缀
+        executor.initialize();
+        return executor;
+    }
+    @Override
+    public AsyncUncaughtExceptionHandler getAsyncUncaughtExceptionHandler() {
+        return new AsyncUncaughtExceptionHandler() {
+            @Override
+            public void handleUncaughtException(Throwable ex, Method method, Object... params) {
+                log.error("Async method {} execution failed: ", method.getName(), ex);
+            }
+        };
+    }
+    @Bean("notifyEventExecutor")
+    public Executor notifyEventExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        // 核心线程数
+        executor.setCorePoolSize(4);
+        // 最大线程数
+        executor.setMaxPoolSize(8);
+        // 队列容量
+        executor.setQueueCapacity(100);
+        // 线程前缀名
+        executor.setThreadNamePrefix("notify-event-");
+        // 线程存活时间
+        executor.setKeepAliveSeconds(60);
+        // 拒绝策略：由调用线程处理
+        executor.setRejectedExecutionHandler(new ThreadPoolExecutor.CallerRunsPolicy());
+        // 等待所有任务完成后再关闭线程池
+        executor.setWaitForTasksToCompleteOnShutdown(true);
+        executor.initialize();
+        return executor;
+    }
+
+}
